v3.0
added the possibility to run in series the transformation of the
subboxes from xyz to radecz

changed the case when we want to convert ic

in generate_lightcone when the conversion is done from r to z, sometimes
it can have an issue to convert the number 0. this can be fixed if we
force r to have float64  or by removing the 0 from the array. we select
first the necessary r and the we convert them in this way we remove 0.